requirements.txt
grpcio
grpcio-tools
transformers==4.34.0
torch>=2.0.0
